pyproject.toml
[tool.poetry]
name = "pitgpt"
version = "0.1.0"
description = ""
authors = ["0x7E20A000 <0x7E20@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.10"
pyqt5 = "^5.15.11"
openai = "^1.54.3"
emoji = "^2.14.0"
termcolor = "^2.5.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
